+++ memote/support/consistency.py
@@ -26,7 +26,6 @@ import numpy as np
 from cobra.exceptions import Infeasible
 from cobra.flux_analysis import flux_variability_analysis
 
-import memote.support.helpers as helpers
 import memote.support.consistency_helpers as con_helpers
 
 LOGGER = logging.getLogger(__name__)
@@ -267,12 +266,13 @@ def produce_atp_closed_exchanges(model):
     return state
 
 
-def find_imbalanced_reactions(model):
+def find_mass_imbalanced_reactions(model):
     """
-    Find metabolic reactions that not mass and/or charge balanced.
+    Find metabolic reactions that are not mass balanced.
 
     This will exclude biomass, exchange and demand reactions as they are
-    unbalanced by definition.
+    unbalanced by definition. It will also fail all reactions where at
+    least one metabolite does not have a formula defined.
 
     Parameters
     ----------
@@ -280,11 +280,28 @@ def find_imbalanced_reactions(model):
         The metabolic model under investigation.
 
     """
-    exchanges = set(model.exchanges)
-    biomass = set(helpers.find_biomass_reaction(model))
-    total_rxns = set(model.reactions)
-    metab_rxns = total_rxns - (exchanges | biomass)
-    return [rxn for rxn in metab_rxns if len(rxn.check_mass_balance()) > 0]
+    internal_rxns = con_helpers.get_internals(model)
+    return [
+        rxn for rxn in internal_rxns if not con_helpers.is_mass_balanced(rxn)]
+
+
+def find_charge_imbalanced_reactions(model):
+    """
+    Find metabolic reactions that are not charge balanced.
+
+    This will exclude biomass, exchange and demand reactions as they are
+    unbalanced by definition. It will also fail all reactions where at
+    least one metabolite does not have a charge defined.
+
+    Parameters
+    ----------
+    model : cobra.Model
+        The metabolic model under investigation.
+
+    """
+    internal_rxns = con_helpers.get_internals(model)
+    return [
+        rxn for rxn in internal_rxns if not con_helpers.is_charge_balanced(rxn)]
 
 
 def find_blocked_reactions(model):